Added Wordle and WordLink functionality.
Added code for wordle and wordlink so that both
are playable now. All the txt files titled
allowed-*.txt are lists of words that the wordle
game uses. They were created using the script
build_wordlists.py. They serve as a more efficient
way for the Wordle game to check is a user input
word is a valid word.